Refactor a bit before adding extra pages (#161)
* Refactor a bit before adding extra pages

* Properly implement breadcrumbs

* Fix race condition setting name of an object
