+++ components/layout/base-layout.tsx
@@ -1,7 +1,6 @@
 import type { ReactNodeLike } from 'prop-types';
 import { Layout } from 'antd';
 import NavBar from './header/nav-bar';
-import Breadcrumb from './main/breadcrumb';
 import Footer from './footer/footer';
 
 const { Content } = Layout;
@@ -14,16 +13,7 @@ interface Props {
 const BaseLayout = ({ children, showSearch }: Props) => (
   <Layout style={{ minHeight: '100vh' }}>
     <NavBar showSearch={showSearch} />
-    <Content className="site-layout-main">
-      <Breadcrumb />
-      <Layout
-        hasSider
-        className="site-layout-background"
-        style={{ padding: '24px 0', flexDirection: 'row' }}
-      >
-        {children}
-      </Layout>
-    </Content>
+    <Content>{children}</Content>
     <Footer />
   </Layout>
 );
