+++ components/pages/home/home-page.tsx
@@ -1,0 +1,10 @@
+import BaseLayout from '@/components/layout/base-layout';
+import React from 'react';
+
+const HomePage = () => (
+  <BaseLayout showSearch={false}>
+    <div style={{ padding: '0 24px' }}>WIP...</div>
+  </BaseLayout>
+);
+
+export default HomePage;
