+++ core/hooks/use-top-level-route.tsx
@@ -1,0 +1,9 @@
+import { useRouter } from 'next/router';
+
+export function useTopLevelRoute() {
+  const { asPath } = useRouter();
+
+  const parts = asPath.replace(/^\//, '').split('/');
+
+  return `/${parts[0]}`;
+}
